+++ CLI/main.cpp
@@ -1,6 +1,9 @@
 #include <iostream>
 #include <map>
+#include <sstream>
+#include <fstream>
 
+#include "Utils.hpp"
 #include "Command.hpp"
 #include "OrderedMap.hpp"
 #include "commands/Build.hpp"
@@ -9,6 +12,7 @@
 #include "commands/Dependencies.hpp"
 #include "commands/Help.hpp"
 #include "commands/LinkProject.hpp"
+#include "commands/Rebuild.hpp"
 #include "commands/RunProject.hpp"
 #include "commands/Version.hpp"
 
@@ -20,6 +24,7 @@ static OrderedMap<std::string, std::unique_ptr<Command>> CreateEngineCommands() 
   cmds.insert(COMMAND_CREATE_PROJECT, std::make_unique<CreateProject>());
   cmds.insert(COMMAND_LINK_PROJECT, std::make_unique<LinkProject>());
   cmds.insert(COMMAND_VERSION, std::make_unique<Version>());
+  cmds.insert(COMMAND_REBUILD, std::make_unique<Rebuild>());
   return cmds;
 }
 
@@ -32,7 +37,6 @@ static OrderedMap<std::string, std::unique_ptr<Command>> CreateProjectCommands()
   return cmds;
 }
 
-// TODO: add a file in GameEngine with expected version of cli and if different auto rebuild
 int main(const int argc, const char *argv[]) {
   #if !defined(_WIN32) && !(defined(__APPLE__) && defined(__MACH__)) && !defined(__linux__)
     #error "Unsupported platform. The engine only supports MacOS, Linux and Windows for now."
@@ -51,14 +55,40 @@ int main(const int argc, const char *argv[]) {
     return 0;
   }
 
-  std::vector<std::string> args(argv, argv + argc);
-  args.erase(args.begin());
-  if (commands.contains(args[0])) {
-    commands[args[0]]->run(args, commands);
-    return 1;
+  // Validate all command args start with "--"
+  if (std::string(argv[1]).rfind("--", 0) != 0) {
+    std::cout << RED << "Expected command to start with '--'. Got: '" << argv[1] << "'\n";
+    std::cout << "Try " << CLI_EXECUTE_COMMAND(COMMAND_HELP) << " for usage." << RESET << std::endl;
+    return 0;
   }
 
-  std::cout << RED << "Command " << args[0] << " not found." << "\n";
-  std::cout << "Try '" << CLI_COMMAND << " help' for available commands." << RESET << std::endl;
-  return 0;
+
+  std::vector<std::string> argvStr(argv, argv + argc);
+  argvStr.erase(argvStr.begin());
+  std::string result = Utils::Join(argvStr, " ");
+
+  std::vector<std::string> userCommands = Utils::Split(result, "--");
+  std::vector<std::vector<std::string>> commandArgs;
+  for (auto it = userCommands.begin(); it != userCommands.end();) {
+    *it = Utils::Trim(*it);
+    if (it->empty())
+      it = userCommands.erase(it);
+    else {
+      commandArgs.emplace_back(Utils::Split(*it, " "));
+      ++it;
+    }
+  }
+
+  for (const auto &args: commandArgs) {
+    if (!commands.contains(args[0])) {
+      std::cout << RED << "Command " << args[0] << " not found." << "\n";
+      std::cout << "Try '" << CLI_COMMAND() << " help' for available commands." << RESET << std::endl;
+      return 0;
+    }
+  }
+
+  for (const auto &args: commandArgs)
+    commands[args[0]]->run(args, commands);
+
+  return 1;
 }